Align navigation with null stack products
