+++ CLAUDE.md
@@ -1,6 +1,5 @@
 # Documentation
 
-Project documentation has moved to [README.md](README.md) and `docs/`.
+Project documentation lives in [README.md](README.md) and `docs/`.
 
 - Read first: `README.md`
-- Design docs: `docs/plans/`
