+++ ui/src/routes/instances/[component]/[name]/+page.svelte
@@ -13,7 +13,8 @@
   import NullTicketsPanel from "$lib/components/NullTicketsPanel.svelte";
   import { api, type ApiRequestError } from "$lib/api/client";
   import {
-    orchestrationUiRoutes,
+    nullboilerUiRoutes,
+    nullticketsUiRoutes,
     withBoilerInstance,
     withTicketsInstance,
   } from "$lib/orchestration/routes";
@@ -630,7 +631,7 @@
   async function openTicketsStore() {
     if (component !== "nulltickets") return;
     setSelectedTicketsInstance(name);
-    await goto(withTicketsInstance(orchestrationUiRoutes.store(), name));
+    await goto(withTicketsInstance(nullticketsUiRoutes.store(), name));
   }
 
   async function createTicketTask() {
@@ -1048,7 +1049,7 @@
       <button class="btn" onclick={stop} disabled={loading}>Stop</button>
       <button class="btn" onclick={restart} disabled={loading}>Restart</button>
       {#if component === "nullwatch"}
-        <a class="btn" href={`/observability?watch=${encodeURIComponent(name)}`}>Observability</a>
+        <a class="btn" href={`/nullwatch?watch=${encodeURIComponent(name)}`}>NullWatch</a>
       {/if}
       <button class="btn danger" onclick={remove} disabled={loading}
         >Delete</button
@@ -1285,8 +1286,8 @@
                   {#if linkedWatch}
                     <a
                       class="btn integration-btn"
-                      href={`/observability?watch=${encodeURIComponent(linkedWatch.name)}`}
-                      >Open Observability</a
+                      href={`/nullwatch?watch=${encodeURIComponent(linkedWatch.name)}`}
+                      >Open NullWatch</a
                     >
                   {/if}
                 </div>
@@ -1339,8 +1340,8 @@
                   </button>
                   <a
                     class="btn integration-btn"
-                    href={`/observability?watch=${encodeURIComponent(name)}`}
-                    >Open Observability</a
+                    href={`/nullwatch?watch=${encodeURIComponent(name)}`}
+                    >Open NullWatch</a
                   >
                 </div>
               {:else}
@@ -1447,13 +1448,13 @@
               <div class="integration-actions">
                 <button
                   class="btn integration-btn"
-                  onclick={() => openBoilerRoute(orchestrationUiRoutes.workflows())}
+                  onclick={() => openBoilerRoute(nullboilerUiRoutes.workflows())}
                 >
                   Workflows
                 </button>
                 <button
                   class="btn integration-btn"
-                  onclick={() => openBoilerRoute(orchestrationUiRoutes.runs())}
+                  onclick={() => openBoilerRoute(nullboilerUiRoutes.runs())}
                 >
                   Runs
                 </button>
